Original average.py submission
I originally submitted this file for grading. This program makes use of
the sum() function in Python to sum the list of numbers. However, before being
graded I decided that the point of the lesson was to show accumulator
principles in the design of the program. I then submitted average2.py
for grading.