Fixed Multiplication
We just add one zero each time because we are cascading the outputs.
This effectively shifts it by 1 zero at each level. Before it was not
working because we were shifting by an increasing amount, which is not
how multiplication works and we were losing our entire number by the
time we shift off by 31.