Use absolute value for bearing
Not using PID for distance, however need to still tweak final constant values for bearing PID controller, this push fixes the search so we are using absolute value of bearing: before we would get negative results which would cause it to prematurely exit the search mode. Now the bot will properly turn left or right to adjust, but at incorrect speeds.